feat: move service methods to separate file
